added websocket close tests; refactored some of the other tests
